pyproject.toml
[tool.poetry]
name = "gitlab-ticker"
version = "0.1.0"
description = ""
authors = ["Pierre Rossinès <pierre.rossines@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^1.0.3"
langchain-anthropic = "^1.0.1"
python-dotenv = "^1.2.1"
langchain-openai = "^1.0.2"
langsmith = "^0.4.41"
slack-sdk = "^3.27.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.3"
mypy = "^1.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Ruff configuration
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["gitlab_ticker"]

[tool.mypy]
# Mypy configuration
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "slack_sdk.*",
]
ignore_missing_imports = true
